refactor: Add clear map marker layer before rendering new markers.
