+++ src/components/map/map.tsx
@@ -3,12 +3,10 @@ import { OfferPreview } from '../../types/offer';
 import { City } from '../../types/city';
 import { Page } from '../../types/page';
 import { getMapClasses } from './map-utils';
+import { defaultCustomIcon, activeCustomIcon } from './pin-icons';
 import useMap from '../../hooks/use-map';
-import leaflet from 'leaflet';
+import leaflet, { LayerGroup } from 'leaflet';
 import 'leaflet/dist/leaflet.css';
-
-const URL_PIN_DEFAULT = '../../../public/img/pin.svg';
-const URL_PIN_ACTIVE = '../../../public/img/pin-active.svg';
 
 type MapProps = {
   pageType: Page;
@@ -19,24 +17,22 @@ type MapProps = {
 
 function Map(props: MapProps): JSX.Element {
   const { pageType, city, offerPreviews, hoveredOffer } = props;
+
   const mapRef = useRef(null);
   const map = useMap(mapRef, city);
   const mapClasses = getMapClasses(pageType);
 
+  const markerLayerRef = useRef<LayerGroup>(leaflet.layerGroup());
+
   useEffect(() => {
-    const defaultCustomIcon = leaflet.icon({
-      iconUrl: URL_PIN_DEFAULT,
-      iconSize: [40, 40],
-      iconAnchor: [20, 40],
-    });
-
-    const activeCustomIcon = leaflet.icon({
-      iconUrl: URL_PIN_ACTIVE,
-      iconSize: [40, 40],
-      iconAnchor: [20, 40],
-    });
-
     if (map) {
+      markerLayerRef.current.addTo(map);
+    }
+  }, [map]);
+
+  useEffect(() => {
+    if (map) {
+      markerLayerRef.current.clearLayers();
       offerPreviews.forEach((offerPreview) => {
         leaflet
           .marker(
@@ -51,7 +47,7 @@ function Map(props: MapProps): JSX.Element {
                   : defaultCustomIcon,
             }
           )
-          .addTo(map);
+          .addTo(markerLayerRef.current);
       });
     }
   }, [map, offerPreviews, hoveredOffer]);
